Refactor: 익셉션 패키징
