+++ src/test/java/chap07/UserRegisterTest.java
@@ -3,6 +3,8 @@ package chap07;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 
+import chap07.exception.DupIdException;
+import chap07.exception.WeakPasswordException;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
