+++ src/test/java/chap07/exception/DupIdException.java
@@ -1,4 +1,4 @@
-package chap07;
+package chap07.exception;
 
 public class DupIdException extends RuntimeException{
