+++ src/test/java/chap07/exception/WeakPasswordException.java
@@ -1,4 +1,4 @@
-package chap07;
+package chap07.exception;
 
 public class WeakPasswordException extends RuntimeException{
